V83f4
merged harmony patches in one file, reordered, cleaned
