+++ Source/1.4/RimworldMod.csproj
@@ -112,6 +112,7 @@
     <Compile Include="ArrivalAction\CaravanArrivalAction_VisitInsectPillarSite.cs" />
     <Compile Include="ArrivalAction\CaravanArrivalAction_VisitImpactSite.cs" />
     <Compile Include="ArrivalAction\TransportPodsArrivalAction_ShipAssault.cs" />
+    <Compile Include="HarmonyPatches.cs" />
     <Compile Include="Jobs\JobDriver_LoadShipBuilding.cs" />
     <Compile Include="Jobs\JobGiver_LoadShipBuilding.cs" />
     <Compile Include="Quest\QuestNode_RoyalAscentShip.cs" />
@@ -322,7 +323,6 @@
     <Compile Include="ShipHeatNet.cs" />
     <Compile Include="ShipMoveBlueprint.cs" />
     <Compile Include="ShuttleCaravanUtility.cs" />
-    <Compile Include="ShuttleMod.cs" />
     <Compile Include="SpaceSite.cs" />
     <Compile Include="MapGen\SymbolResolver_DebrisEdgeStreet.cs" />
     <Compile Include="MapGen\SymbolResolver_DebrisClump.cs" />
